+++ src/js/pages/lines/linesTemplate.js
