+++ src/js/pages/main/templates/appNavbar.js
